+++ CHANGELOG.md
@@ -4,4 +4,5 @@ Each change log are assumed to be accompanied by associated unit test and docume
 
 ## 0.5.0
 
+- added ability to be able to put free form new indicator in with the list of options (#2)
 - initial release with change log file